Migrate consul container test to integration folder
