+++ server/src/integration/resources/config/consul.yml
