+++ server/src/test/java/com/networknt/server/RegistryTest.java
@@ -16,172 +16,99 @@
 
 package com.networknt.server;
 
-import com.networknt.cluster.LightCluster;
 import com.networknt.config.Config;
-import com.networknt.consul.ConsulConfig;
-import com.networknt.consul.ConsulRegistry;
-import com.networknt.consul.client.ConsulClientImpl;
-import com.networknt.registry.URL;
-import com.networknt.registry.URLImpl;
-import com.networknt.utility.Constants;
-import com.networknt.utility.Util;
 import org.junit.*;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import org.testcontainers.containers.GenericContainer;
 
-import java.net.InetAddress;
+import java.io.File;
+import java.lang.reflect.Method;
+import java.net.URL;
+import java.net.URLClassLoader;
 import java.util.HashMap;
 import java.util.Map;
-import java.util.concurrent.CountDownLatch;
-import java.util.concurrent.TimeUnit;
-
-import static com.networknt.consul.ConsulConstants.CONFIG_NAME;
-import static com.networknt.server.Server.STATUS_HOST_IP;
-import static org.junit.Assert.fail;
 
 /**
- * This test is an automatic test for real consul registry and discovery by using testContainers.
- * Since it depends on docker so it should be disabled all the time unless it is used.
+ * This test depends on consul server so it should be disabled all the time unless
+ * it is used. To start consul use the following command line.
+ *
+ * docker run -d -p 8400:8400 -p 8500:8500/tcp -p 8600:53/udp -e 'CONSUL_LOCAL_CONFIG={"acl_datacenter":"dc1","acl_default_policy":"deny","acl_down_policy":"extend-cache","acl_master_token":"the_one_ring","bootstrap_expect":1,"datacenter":"dc1","data_dir":"/usr/local/bin/consul.d/data","server":true}' consul agent -server -ui -bind=127.0.0.1 -client=0.0.0.0
+ *
+ * To access consul ui in order to check registered services, use this url: http://localhost:8500
+ *
+ * Created by steve on 29/01/17.
  */
-@Ignore
 public class RegistryTest {
     static final Logger logger = LoggerFactory.getLogger(RegistryTest.class);
-    static ConsulRegistry registry;
-    static LightCluster lightCluster;
-    static String serviceId = "com.networknt.petstore-1.0.0";
-    static String ipAddress;
-    static CountDownLatch latch = new CountDownLatch(1);
+    static final String homeDir = System.getProperty("user.home");
 
-    static ConsulConfig consulConfig = (ConsulConfig) Config.getInstance().getJsonObjectConfig(CONFIG_NAME, ConsulConfig.class);
-    static ServerConfig serverConfig = (ServerConfig) Config.getInstance().getJsonObjectConfig("server", ServerConfig.class);
+    static Server server = null;
 
-    static Server server1 = null;
-    static Server server2 = null;
-    static Server server3 = null;
+    //@BeforeClass
+    public static void setUp() throws Exception {
+        // inject server config here.
+        Config config = Config.getInstance();
+        // write a config file into the user home directory
+        Map<String, Object> map = new HashMap<String, Object>();
+        map.put("description", "server config");
+        map.put("enableHttp", true);
+        map.put("ip", "0.0.0.0");
+        map.put("httpPort", 8080);
+        map.put("enableHttps", true);
+        map.put("httpsPort", 8443);
+        map.put("keystoreName", "server.keystore");
+        map.put("keystorePass", "secret");
+        map.put("keyPass", "secret");
+        map.put("serviceId", "com.networknt.apia-1.0.0");
+        map.put("enableRegistry", true);
+        addURL(new File(homeDir).toURI().toURL());
+        config.getMapper().writeValue(new File(homeDir + "/server.json"), map);
 
-    @ClassRule
-    public static GenericContainer consul
-            = new GenericContainer("consul")
-            .withExposedPorts(8500);
-
-    @BeforeClass
-    public static void setUp() throws InterruptedException {
-        //set ip address
-        ipAddress = System.getenv(STATUS_HOST_IP);
-        logger.info("Registry IP from STATUS_HOST_IP is " + ipAddress);
-        if (ipAddress == null) {
-            InetAddress inetAddress = Util.getInetAddress();
-            ipAddress = inetAddress.getHostAddress();
-            logger.info("Could not find IP from STATUS_HOST_IP, use the InetAddress " + ipAddress);
+        if (server == null) {
+            logger.info("starting server");
+            Server.start();
         }
-        //refresh consul config with dynamic port generated by container test
-        Integer port = consul.getFirstMappedPort();
-        Map<String, String> params = new HashMap<>();
-        params.put("registrySessionTimeout", "1000");
-        consulConfig.setConsulUrl("http://localhost:" + port);
-        //construct registry
-        registry = new ConsulRegistry(new URLImpl("light", "localhost", port, "", params), new ConsulClientImpl());
-        //construct cluster
-        lightCluster = new LightCluster();
-        //start server without registry, the registry is be done separately for testing
-        startServer(49588, null, server1);
-        startServer(49589, "", server2);
-        startServer(49590, "dev", server3);
-        //wait for registry complete
-        latch.await(10, TimeUnit.SECONDS);
     }
 
-    @AfterClass
-    public static void tearDown() {
-        if (server1 != null) {
+    //@AfterClass
+    public static void tearDown() throws Exception {
+        if (server != null) {
             try {
                 Thread.sleep(1000);
             } catch (InterruptedException ignored) {
 
             }
             Server.stop();
-            logger.info("The server1 is stopped.");
-        }
-        if (server2 != null) {
-            try {
-                Thread.sleep(1000);
-            } catch (InterruptedException ignored) {
-
+            logger.info("The server is stopped.");
             }
-            server2.stop();
-            logger.info("The server2 is stopped.");
         }
-        if (server3 != null) {
-            try {
-                Thread.sleep(1000);
-            } catch (InterruptedException ignored) {
 
-            }
-            server3.stop();
-            logger.info("The server3 is stopped.");
-        }
+    public static void addURL(URL url) throws Exception {
+        URLClassLoader classLoader
+                = (URLClassLoader) ClassLoader.getSystemClassLoader();
+        Class clazz = URLClassLoader.class;
+
+        // Use reflection
+        Method method = clazz.getDeclaredMethod("addURL", new Class[]{URL.class});
+        method.setAccessible(true);
+        method.invoke(classLoader, new Object[]{url});
     }
 
     @Test
-    public void testDiscoverWithoutEnvTag() {
-        String url = lightCluster.serviceToUrl("https", serviceId, null, null);
-        Assert.assertNotNull(url);
-        Assert.assertTrue(url.equals("https://" + ipAddress + ":49588") || url.equals("https://" + ipAddress + ":49589"));
+    public void testVoid() {
+
     }
 
     @Test
-    public void testDiscoverWithEmptyStringEnvTag() {
-        String url = lightCluster.serviceToUrl("https", serviceId, "", null);
-        Assert.assertNotNull(url);
-        Assert.assertTrue(url.equals("https://" + ipAddress + ":49588") || url.equals("https://" + ipAddress + ":49589"));
-    }
-
-    @Test
-    public void testDiscoverWithNonEmptyStringEnvTag() {
-        String url = lightCluster.serviceToUrl("https", serviceId, "dev", null);
-        Assert.assertNotNull(url);
-        Assert.assertEquals("https://" + ipAddress + ":49590", url);
-    }
-
-    @Test
-    public void testDiscoverWithNonexistentEnvTag() {
+    @Ignore
+    public void testServer() {
+        // server cannot be started as there is no spi routing handler provider
+        Assert.assertNull(server);
         try {
-            String url = lightCluster.serviceToUrl("https", serviceId, "sit", null);
-            fail();
-        } catch(Exception e) {
+            Thread.sleep(60000);
+        } catch (InterruptedException ignored) {
 
         }
     }
 
-    @Test
-    public void testDiscoverCaching() {
-        String url = lightCluster.serviceToUrl("https", serviceId, "dev", null);
-        Assert.assertNotNull(url);
-        Assert.assertEquals("https://" + ipAddress + ":49590", url);
-        try {
-            lightCluster.serviceToUrl("https", serviceId, "sit", null);
-            fail();
-        } catch(Exception e) {
-
-        }
-        url = lightCluster.serviceToUrl("https", serviceId, "dev", null);
-        Assert.assertNotNull(url);
-        Assert.assertEquals("https://" + ipAddress + ":49590", url);
-    }
-
-    private static void startServer(int port, String envTag, Server server) {
-        serverConfig.setHttpsPort(port);
-        if (server2 == null) {
-            logger.info("starting server2");
-            server2.start();
-        }
-        URL serviceUrl = new URLImpl("https", ipAddress, port, serviceId, new HashMap<>());
-        if (envTag == null) {
-            serviceUrl.removeParameter(Constants.TAG_ENVIRONMENT);
-        } else {
-            serviceUrl.addParameter(Constants.TAG_ENVIRONMENT, envTag);
-        }
-        registry.register(serviceUrl);
-    }
 }
